pyproject.toml
[tool.poetry]
name = "sibdevtask"
version = "0.1.0"
description = ""
authors = ["Aybek18 <kasymaliev09@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
django-environ = "^0.10.0"
drf-spectacular = "^0.26.4"
gunicorn = "^21.2.0"
django-redis = "^5.3.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
black = "^23.1.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
